refactor: 프롬프트 효율성 체크 모델 분리
